storetestcases: move evidencestore and keyvaluestore tests to fixtures
