--- store/storetestcases/deletevalue.go
@@ -1,126 +1,0 @@
-// Copyright 2017 Stratumn SAS. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package storetestcases
-
-import (
-	"fmt"
-	"io/ioutil"
-	"log"
-	"strings"
-	"sync/atomic"
-	"testing"
-
-	"github.com/stratumn/sdk/testutil"
-	"github.com/stretchr/testify/assert"
-)
-
-// TestDeleteValue tests what happens when you delete an existing value.
-func (f Factory) TestDeleteValue(t *testing.T) {
-	a := f.initKeyValueStore(t)
-	defer f.freeKeyValueStore(a)
-
-	key := testutil.RandomKey()
-	value1 := testutil.RandomValue()
-	a.SetValue(key, value1)
-
-	value2, err := a.DeleteValue(key)
-	assert.NoError(t, err, "a.DeleteValue()")
-	assert.EqualValues(t, value1, value2, "a.DeleteValue() should return the deleted value")
-
-	value2, err = a.GetValue(key)
-	assert.NoError(t, err, "a.GetValue()")
-	assert.Nil(t, value2, "Deleted value should not be found")
-}
-
-// TestDeleteValueNotFound tests what happens when you delete a nonexistent
-// value.
-func (f Factory) TestDeleteValueNotFound(t *testing.T) {
-	a := f.initKeyValueStore(t)
-	defer f.freeKeyValueStore(a)
-
-	v, err := a.DeleteValue(testutil.RandomKey())
-	assert.NoError(t, err, "a.DeleteValue()")
-	assert.Nil(t, v, "Not found value should be nil")
-}
-
-func searchNewKey(values map[string][]byte) ([]byte, string) {
-	for {
-		k := testutil.RandomKey()
-		strkey := strings.Trim(strings.Join(strings.Fields(fmt.Sprint(k)), ""), "[]")
-		if _, present := values[strkey]; !present {
-			return k, strkey
-		}
-	}
-}
-
-// BenchmarkDeleteValue benchmarks deleting existing segments.
-func (f Factory) BenchmarkDeleteValue(b *testing.B) {
-	a := f.initKeyValueStoreB(b)
-	defer f.freeKeyValueStore(a)
-
-	values := make(map[string][]byte, b.N)
-	for i := 0; i < b.N; i++ {
-		k, strkey := searchNewKey(values)
-		v := testutil.RandomValue()
-		a.SetValue(k, v)
-		values[strkey] = k
-	}
-
-	b.ResetTimer()
-	log.SetOutput(ioutil.Discard)
-
-	for _, k := range values {
-		if s, err := a.DeleteValue(k); err != nil {
-			b.Error(err)
-		} else if s == nil {
-			b.Error("s = nil want []byte")
-		}
-	}
-}
-
-// BenchmarkDeleteValueParallel benchmarks deleting existing segments in
-// parallel.
-func (f Factory) BenchmarkDeleteValueParallel(b *testing.B) {
-	a := f.initKeyValueStoreB(b)
-	defer f.freeKeyValueStore(a)
-
-	mapvalues := make(map[string][]byte, b.N)
-	for i := 0; i < b.N; i++ {
-		k, strkey := searchNewKey(mapvalues)
-		v := testutil.RandomValue()
-		a.SetValue(k, v)
-		mapvalues[strkey] = k
-	}
-	values := make([][]byte, 0, b.N)
-	for _, v := range mapvalues {
-		values = append(values, v)
-	}
-
-	var counter uint64
-
-	b.ResetTimer()
-	log.SetOutput(ioutil.Discard)
-
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			i := atomic.AddUint64(&counter, 1) - 1
-			if s, err := a.DeleteValue(values[i]); err != nil {
-				b.Error(err)
-			} else if s == nil {
-				b.Error("s = nil want *cs.Segment")
-			}
-		}
-	})
-}
